# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Our site
  version: 0.1.0
paths:
...
  /api/users/me/favourites:
    get:
      tags:
      - Favourites
      summary: Получить список избранного текущего пользователя
      operationId: get_favourites_me_api_users_me_favourites_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    post:
      tags:
      - Favourites
      summary: Добавить товар в избранное текущего пользователя
      operationId: add_to_favourites_me_api_users_me_favourites_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavouriteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
    delete:
      tags:
      - Favourites
      summary: Очистить всё избранное текущего пользователя
      operationId: clear_favourites_me_api_users_me_favourites_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/favourites:
    get:
      tags:
      - Favourites
      summary: Получить список избранного пользователя по ID (Admin)
      operationId: get_favourites_by_user_id_api_users__user_id__favourites_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    post:
      tags:
      - Favourites
      summary: Добавить товар в избранное пользователя по ID (Admin)
      operationId: add_to_favourites_by_user_id_api_users__user_id__favourites_post
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavouriteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
    delete:
      tags:
      - Favourites
      summary: Очистить всё избранное пользователя по ID (Admin)
      operationId: clear_favourites_by_user_id_api_users__user_id__favourites_delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
  /api/users/me/favourites/{product_id}:
    delete:
      tags:
      - Favourites
      summary: Удалить товар из избранного текущего пользователя
      operationId: remove_from_favourites_me_api_users_me_favourites__product_id__delete
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - AUTHENTICATED
      description: '**Требуемые роли:** Любой авторизованный пользователь'
  /api/users/{user_id}/favourites/{product_id}:
    delete:
      tags:
      - Favourites
      summary: Удалить товар из избранного пользователя по ID (Admin)
      operationId: remove_from_favourites_by_user_id_api_users__user_id__favourites__product_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouritesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Пользователь не аутентифицирован.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
        '403':
          description: Недостаточно прав доступа для выполнения данной операции.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailResponse'
      security:
      - BearerAuth: []
      x-required-roles:
      - ADMIN
      description: '**Требуемые роли:** `ADMIN`'
...
components:
  schemas:
...
    AddFavouriteRequest:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
          description: ID добавляемого товара
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
          description: Заметка к товару в избранном
      type: object
      required:
      - product_id
      title: AddFavouriteRequest
...
    FavouriteItemResponse:
      properties:
        product_id:
          type: string
          format: uuid
          title: Product Id
        added_user_id:
          type: string
          format: uuid
          title: Added User Id
        updated_at:
          type: string
          format: date-time
          title: Updated At
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - product_id
      - added_user_id
      - updated_at
      title: FavouriteItemResponse
    FavouritesResponse:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        products:
          items:
            $ref: '#/components/schemas/FavouriteItemResponse'
          type: array
          title: Products
        total_count:
          type: integer
          title: Total Count
      type: object
      required:
      - user_id
      - products
      - total_count
      title: FavouritesResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ErrorDetailResponse:
      title: ErrorDetailResponse
      type: object
      properties:
        detail:
          title: Detail
          type: string
          description: Описание ошибки
      required:
      - detail
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Access Token (в заголовке `Authorization: Bearer <token>`
        или cookie `access_token`)'